moved LazyGyroscope to correct package
