+++ src/org/ljrobotics/lib/util/drivers/LazyGyroscope.java
@@ -1,4 +1,4 @@
-package org.ljrobotics.frc2018.utils;
+package org.ljrobotics.lib.util.drivers;
 
 import edu.wpi.first.wpilibj.ADXRS450_Gyro;
